Refactor logic to new directory
